main: use unique exitcode for incorrect invocation
Command line utilities typically exit with a specific exitcode to
indicate incorrect invocation. Golang has settled on 2, so let's go with
that.

This also moves the file extension check before the check for the
firmware's existence. No point in validating that the file exists if we
already know it has the wrong extension.